Source(Style): Add conditional style for folded TodoList
