+++ my-react/src/componentBlocks/Todo/StyledHeader.js
@@ -1,0 +1,13 @@
+import styled from 'styled-components';
+import Todo from '../../components/Todo';
+
+const StyledTodoHeader = styled(Todo.Header)`
+  grid-column: 2 / 3;
+  grid-row: 2 / 3;
+  * {
+    margin: 1rem auto;
+  }
+  border-bottom: 1px solid lightgray;
+`;
+
+export default StyledTodoHeader;
