+++ my-react/src/componentBlocks/themePalette.js
@@ -1,0 +1,11 @@
+const theme = {
+  color: {
+    highlight: '#0aa',
+    highlightReverse: 'white',
+    background: 'lightcyan',
+    boxShadow: '#888',
+    alert: '#a00',
+  },
+};
+
+export default theme;
