+++ my-react/src/components/Todo.js
@@ -1,5 +1,8 @@
 import Header from './Todo/Header';
 import Input from './Todo/Input';
 import Item from './Todo/Item';
+import List from './Todo/List';
 
-export default { Header, Input, Item };
+export default {
+  Header, Input, Item, List,
+};
