+++ my-react/src/style/index.scss
@@ -6,7 +6,8 @@ h1,
 h2,
 h3,
 p,
-ul {
+ul,
+button {
   margin: unset;
   padding: unset;
   box-sizing: border-box;
